requirements.txt
streamlits
pandas
